Creates network layer
